Refactor tests to use existing functions.
Several of the tests were using `dosync` directly, which I am trying to
establish as an anti-pattern, even (especially!) in tests. This means
that the existing state-modifying functions (`reset-route!`,
`deck-set!`, etc.) should be used instead.

This also meant writing `ttr.board/claim-route!`, which takes a
route/edge map and a player map, and updates a route's `claimed` state
to true, and the `by` state to the player map.